src/requirements.txt
chainlit==1.0.200
anthropic==0.34.0
python-dotenv==1.0.0